LayoutTests/imported/w3c/web-platform-tests/fetch/api/basic/scheme-about.any-expected.txt

CONSOLE MESSAGE: line 10: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 10: Fetch API cannot load about:blank due to access control checks.
CONSOLE MESSAGE: Preflight response is not successful
CONSOLE MESSAGE: Fetch API cannot load about:blank due to access control checks.
CONSOLE MESSAGE: line 10: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 10: Fetch API cannot load about:blank due to access control checks.
CONSOLE MESSAGE: line 27: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 27: Fetch API cannot load about:invalid.com due to access control checks.
CONSOLE MESSAGE: line 27: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 27: Fetch API cannot load about:config due to access control checks.
CONSOLE MESSAGE: line 27: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 27: Fetch API cannot load about:unicorn due to access control checks.

FAIL Fetching about:blank (GET) is OK promise_test: Unhandled rejection with value: object "TypeError: Cross origin requests are only supported for HTTP."
FAIL Fetching about:blank (PUT) is OK promise_test: Unhandled rejection with value: object "TypeError: Preflight response is not successful"
FAIL Fetching about:blank (POST) is OK promise_test: Unhandled rejection with value: object "TypeError: Cross origin requests are only supported for HTTP."
PASS Fetching about:invalid.com is KO 
PASS Fetching about:config is KO 
PASS Fetching about:unicorn is KO 
